Init Groovy/Scala library with `java-library` plugin applied
It is now always working and recommended to apply the `java-library`
plugin when developing a library in Groovy or Scala. This allows
users to define 'api' dependencies. Before, users had to utilize
'compile' (which is now deprecated). When using 'gradle init' this new
recommendation should be reflected to avoid confusion.
This is also in line with what we do for a Kotlin library created with
'gradle init'.